refactor(bilibili): share relation helpers
